+++ SingularityUI/app/actions/api/taskFiles.es6
@@ -1,0 +1,3 @@
+import buildApiAction from './base';
+
+export const FetchAction = buildApiAction('FETCH_TASK_FILES', (taskId, path) => `/sandbox/${taskId}/browse?path=${path || ''}`);
