+++ SingularityUI/app/components/common/Breadcrumbs.jsx
@@ -8,6 +8,10 @@ export default class Breadcrumbs extends React.Component {
         return (
           <li key={n}>{i.label} <a href={i.link}>{i.text}</a></li>
         );
+      } else if (i.onClick) {
+        return (
+          <li key={n}>{i.label} <a onClick={i.onClick}>{i.text}</a></li>
+        );
       } else {
         return (
           <li key={n}>{i.label} {i.text}</li>
@@ -28,7 +32,7 @@ export default class Breadcrumbs extends React.Component {
 
 Breadcrumbs.propTypes = {
   items: React.PropTypes.arrayOf(React.PropTypes.shape({
-      label: React.PropTypes.string.isRequired,
+      label: React.PropTypes.string,
       text: React.PropTypes.oneOfType([React.PropTypes.string.isRequired, React.PropTypes.number.isRequired]),
       link: React.PropTypes.string
   })).isRequired,
